Removed redundancy of attribute models
