+++ filip/models/ngsi_v2/context.py
@@ -15,7 +15,7 @@ from pydantic import \
     root_validator, \
     validator
 
-from filip.models.ngsi_v2.base import EntityPattern, Expression
+from filip.models.ngsi_v2.base import EntityPattern, Expression, BaseAttribute, BaseValueAttribute, BaseNameAttribute
 from filip.models.base import DataType, FiwareRegex
 from filip.models.ngsi_v2.units import validate_unit_data
 
@@ -50,67 +50,7 @@ class GetEntitiesOptions(str, Enum):
                        "except that values are not repeated"
 
 
-# NGSIv2 entity models
-class ContextMetadata(BaseModel):
-    """
-    Context metadata is used in FIWARE NGSI in several places, one of them being
-    an optional part of the attribute value as described above. Similar to
-    attributes, each piece of metadata has.
-
-    Note:
-         In NGSI it is not foreseen that metadata may contain nested metadata.
-    """
-    type: Optional[Union[DataType, str]] = Field(
-        title="metadata type",
-        description="a metadata type, describing the NGSI value type of the "
-                    "metadata value Allowed characters "
-                    "are the ones in the plain ASCII set, except the following "
-                    "ones: control characters, whitespace, &, ?, / and #.",
-        max_length=256,
-        min_length=1,
-        regex=FiwareRegex.standard.value  # Make it FIWARE-Safe
-    )
-    value: Optional[Any] = Field(
-        title="metadata value",
-        description="a metadata value containing the actual metadata"
-    )
-
-    @validator('value', allow_reuse=True)
-    def validate_value(cls, value):
-        assert json.dumps(value), "metadata not serializable"
-        return value
-
-
-class NamedContextMetadata(ContextMetadata):
-    """
-    Model for metadata including a name
-    """
-    name: str = Field(
-        titel="metadata name",
-        description="a metadata name, describing the role of the metadata in "
-                    "the place where it occurs; for example, the metadata name "
-                    "accuracy indicates that the metadata value describes how "
-                    "accurate a given attribute value is. Allowed characters "
-                    "are the ones in the plain ASCII set, except the following "
-                    "ones: control characters, whitespace, &, ?, / and #.",
-        max_length=256,
-        min_length=1,
-        regex=FiwareRegex.standard.value  # Make it FIWARE-Safe
-    )
-
-    @root_validator
-    def validate_data(cls, values):
-        if values.get("name", "").casefold() in ["unit",
-                                                 "unittext",
-                                                 "unitcode"]:
-            values.update(validate_unit_data(values))
-        return values
-
-    def to_context_metadata(self):
-        return {self.name: ContextMetadata(**self.dict())}
-
-
-class ContextAttribute(BaseModel):
+class ContextAttribute(BaseAttribute, BaseValueAttribute):
     """
     Model for an attribute is represented by a JSON object with the following
     syntax:
@@ -138,90 +78,10 @@ class ContextAttribute(BaseModel):
         >>> attr = ContextAttribute(**data)
 
     """
-    type: Union[DataType, str] = Field(
-        default=DataType.TEXT,
-        description="The attribute type represents the NGSI value type of the "
-                    "attribute value. Note that FIWARE NGSI has its own type "
-                    "system for attribute values, so NGSI value types are not "
-                    "the same as JSON types. Allowed characters "
-                    "are the ones in the plain ASCII set, except the following "
-                    "ones: control characters, whitespace, &, ?, / and #.",
-        max_length=256,
-        min_length=1,
-        regex=FiwareRegex.standard.value,  # Make it FIWARE-Safe
-    )
-    value: Optional[Union[Union[float, int, bool, str, List, Dict[str, Any]],
-                          List[Union[float, int, bool, str, List,
-                                     Dict[str, Any]]]]] = Field(
-        default=None,
-        title="Attribute value",
-        description="the actual data"
-    )
-    metadata: Optional[Union[Dict[str, ContextMetadata],
-                             NamedContextMetadata,
-                             List[NamedContextMetadata]]] = Field(
-        default={},
-        title="Metadata",
-        description="optional metadata describing properties of the attribute "
-                    "value like e.g. accuracy, provider, or a timestamp")
-
-    @validator('value')
-    def validate_value_type(cls, value, values):
-        """validator for field 'value'"""
-        type_ = values['type']
-        if value:
-            if type_ == DataType.TEXT:
-                if isinstance(value, list):
-                    return [str(item) for item in value]
-                return str(value)
-            if type_ == DataType.BOOLEAN:
-                if isinstance(value, list):
-                    return [bool(item) for item in value]
-                return bool(value)
-            if type_ in (DataType.NUMBER, DataType.FLOAT):
-                if isinstance(value, list):
-                    return [float(item) for item in value]
-                return float(value)
-            if type_ == DataType.INTEGER:
-                if isinstance(value, list):
-                    return [int(item) for item in value]
-                return int(value)
-            if type_ == DataType.DATETIME:
-                return value
-            if type_ == DataType.ARRAY:
-                if isinstance(value, list):
-                    return value
-                raise TypeError(f"{type(value)} does not match "
-                                f"{DataType.ARRAY}")
-            if type_ == DataType.STRUCTUREDVALUE:
-                value = json.dumps(value)
-                return json.loads(value)
-            else:
-                value = json.dumps(value)
-                return json.loads(value)
-        return value
-
-    @validator('metadata')
-    def validate_metadata_type(cls, value):
-        """validator for field 'metadata'"""
-        if isinstance(value, NamedContextMetadata):
-            value = [value]
-        elif isinstance(value, dict):
-            if all(isinstance(item, ContextMetadata)
-                   for item in value.values()):
-                return value
-            json.dumps(value)
-            return {key: ContextMetadata(**item) for key, item in value.items()}
-        if isinstance(value, list):
-            if all(isinstance(item, NamedContextMetadata) for item in value):
-                return {item.name: ContextMetadata(**item.dict(exclude={
-                    'name'})) for item in value}
-            if all(isinstance(item, Dict) for item in value):
-                return {key: ContextMetadata(**item) for key, item in value}
-        raise TypeError(f"Invalid type {type(value)}")
+    pass
 
 
-class NamedContextAttribute(ContextAttribute):
+class NamedContextAttribute(ContextAttribute, BaseNameAttribute):
     """
     Context attributes are properties of context entities. For example, the
     current speed of a car could be modeled as attribute current_speed of entity
@@ -230,18 +90,7 @@ class NamedContextAttribute(ContextAttribute):
     In the NGSI data model, attributes have an attribute name, an attribute type
     an attribute value and metadata.
     """
-    name: str = Field(
-        titel="Attribute name",
-        description="The attribute name describes what kind of property the "
-                    "attribute value represents of the entity, for example "
-                    "current_speed. Allowed characters "
-                    "are the ones in the plain ASCII set, except the following "
-                    "ones: control characters, whitespace, &, ?, / and #.",
-        max_length=256,
-        min_length=1,
-        regex=FiwareRegex.string_protect.value,
-        # Make it FIWARE-Safe
-    )
+    pass
 
 
 class ContextEntityKeyValues(BaseModel):
