+++ tests/models/test_ngsi_v2_iot.py
@@ -5,8 +5,8 @@ Test module for context broker models
 import unittest
 from typing import List
 
-from filip.models.ngsi_v2.iot import BaseAttribute, DeviceCommand, ServiceGroup, \
-    Device, TransportProtocol
+from filip.models.ngsi_v2.iot import DeviceCommand, ServiceGroup, \
+    Device, TransportProtocol, IoTaBaseAttribute
 
 
 class TestContextv2IoTModels(unittest.TestCase):
@@ -37,16 +37,16 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test if all needed fields, detect all invalid strings
         for string in invalid_strings:
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTaBaseAttribute,
                               name=string, type="name",
                               entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTaBaseAttribute,
                               name="name", type=string,
                               entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTaBaseAttribute,
                               name="name", type="name",
                               entity_name=string, entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTaBaseAttribute,
                               name="name", type="name",
                               entity_name="name", entity_type=string)
 
@@ -69,8 +69,8 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test if all needed fields, do not trow wrong errors
         for string in valid_strings:
-            BaseAttribute(name=string, type=string,
-                          entity_name=string, entity_type=string)
+            IoTaBaseAttribute(name=string, type=string,
+                              entity_name=string, entity_type=string)
             DeviceCommand(name=string, type="name")
             ServiceGroup(entity_type=string, resource="", apikey="")
             Device(device_id="", entity_name=string, entity_type=string,
@@ -78,10 +78,10 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test for the special-string protected field if all strings are blocked
         for string in special_strings:
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTaBaseAttribute,
                               name=string, type="name",
                               entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTaBaseAttribute,
                               name="name", type=string,
                               entity_name="name", entity_type="name")
             self.assertRaises(ValidationError,
@@ -89,8 +89,8 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test for the normal protected field if all strings are allowed
         for string in special_strings:
-            BaseAttribute(name="name", type="name",
-                          entity_name=string, entity_type=string)
+            IoTaBaseAttribute(name="name", type="name",
+                              entity_name=string, entity_type=string)
             ServiceGroup(entity_type=string, resource="", apikey="")
             Device(device_id="", entity_name=string, entity_type=string,
                    transport=TransportProtocol.HTTP)
